+++ CHANGELOG.md
@@ -16,9 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - control system module (part 1):
   - [#967](http://github.com/Nelson-numerical-software/nelson/issues/967) control system module template.
-  - `abcdchk` Verifies the dimensional compatibility of matrices A, B, C, and D.
   - [#944](http://github.com/Nelson-numerical-software/nelson/issues/944) `mag2db`, `db2mag`, `pow2db`, `db2pow` functions.
   - [#968](http://github.com/Nelson-numerical-software/nelson/issues/968) `zp2tf`: Zero-pole to transfer function conversion.
+  - [#954](http://github.com/Nelson-numerical-software/nelson/issues/954) `dcgain`: Low-frequency (DC) gain of LTI system.
+  - `abcdchk` Verifies the dimensional compatibility of matrices A, B, C, and D.
   - `ss`: State-space model.
   - `tf`: Transfer function model (display, horzcat, vertcat, size).
   - `isct`: checks if dynamic system model is in continuous time.
@@ -26,6 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `isstatic`: checks if model is static or dynamic.
   - `islti`: checks if variable is an linear model tf, ss or zpk.
   - `issiso`: checks if dynamic system model is a single input and single output.
+  - `zero`: Zeros and gain of SISO dynamic system.
+  - `pole`: Poles of dynamic system.
 
 ### Changed